convert_descriptor_to_operator abstracted out
